test: extract device scanner setup tests
